+++ client-sdk/ts-web/rt/src/transaction.ts
@@ -123,3 +123,14 @@ export async function signUnverifiedTransaction(
 export type SignatureMessageHandlersWithChainContext = {
     [SIGNATURE_CONTEXT_BASE]?: oasis.signature.MessageHandlerWithChainContext<types.Transaction>;
 };
+
+export type CallHandler<BODY> = (body: BODY) => void;
+export type CallHandlers = {[method: string]: CallHandler<unknown>};
+
+export function visitCall(handlers: CallHandlers, call: types.Call) {
+    if (call.method in handlers) {
+        handlers[call.method](call.body);
+        return true;
+    }
+    return false;
+}